+++ src/openapi/streaming/connection/transport/signalr-core-transport.ts
@@ -63,6 +63,7 @@ class SignalrCoreTransport implements StreamingTransportInterface {
     receivedCallback: ReceiveCallback = NOOP;
     unauthorizedCallback = NOOP;
     setConnectionSlowCallback = NOOP;
+    subscriptionResetCallback = NOOP;
     transportFailCallback;
 
     utf8Decoder!: TextDecoder;
@@ -333,7 +334,7 @@ class SignalrCoreTransport implements StreamingTransportInterface {
                 }
             })
             .catch((error) => {
-                log.error(
+                log.warn(
                     LOG_AREA,
                     'Error occurred while connecting to streaming service',
                     {
@@ -410,9 +411,18 @@ class SignalrCoreTransport implements StreamingTransportInterface {
         this.messageStream = messageStream;
     }
 
+    isNetworkError(error: Error | undefined) {
+        return error?.message?.match(
+            /WebSocket closed with status code: (1006|1011)|Server timeout elapsed without receiving a message|Failed to fetch|Not found|Network request failed|Invocation canceled due to the underlying connection being closed/,
+        );
+    }
+
     handleConnectionClosure(error?: Error) {
         if (error) {
-            log.error(LOG_AREA, 'connection closed abruptly', { error });
+            log.warn(LOG_AREA, 'connection closed abruptly', {
+                error,
+                isNetworkError: this.isNetworkError(error),
+            });
         }
 
         // Do not trigger disconnect in case of transport fallback to avoid reconnection
@@ -455,7 +465,7 @@ class SignalrCoreTransport implements StreamingTransportInterface {
         try {
             const normalizedMessage = this.normalizeMessage(message, protocol);
             const data = this.parseMessage(normalizedMessage, this.utf8Decoder);
-
+            let resetSubscriptions = false;
             if (
                 this.lastMessageId &&
                 data.MessageId !== this.lastMessageId + 1
@@ -474,7 +484,7 @@ class SignalrCoreTransport implements StreamingTransportInterface {
                         messageId: data.MessageId,
                     });
                 } else {
-                    log.error(
+                    log.warn(
                         LOG_AREA,
                         'Missing a streaming signalr-core message',
                         {
@@ -482,7 +492,13 @@ class SignalrCoreTransport implements StreamingTransportInterface {
                             messageId: data.MessageId,
                         },
                     );
+                    resetSubscriptions = true;
                 }
+            }
+
+            if (resetSubscriptions) {
+                this.subscriptionResetCallback();
+                return;
             }
 
             this.lastMessageId = data.MessageId as number;
@@ -503,17 +519,18 @@ class SignalrCoreTransport implements StreamingTransportInterface {
         }
     }
 
-    handleMessageStreamError(error: unknown) {
+    handleMessageStreamError(error: Error | undefined) {
         // It will be called if signalr failed to send message to start streaming
         // or if connection is closed with some error
         // only handle the 1st case since connection closing with error is already handled in onclose handler
         // It will trigger disconnected state and will eventually try to reconnect again
         if (!this.hasStreamingStarted) {
-            log.error(
+            log.warn(
                 LOG_AREA,
                 'Error occurred while starting message streaming',
                 {
                     error,
+                    isNetworkError: this.isNetworkError(error),
                 },
             );
 
@@ -647,6 +664,10 @@ class SignalrCoreTransport implements StreamingTransportInterface {
 
     setUnauthorizedCallback(callback: () => void) {
         this.unauthorizedCallback = callback;
+    }
+
+    setSubscriptionResetCallback(callback: () => void) {
+        this.subscriptionResetCallback = callback;
     }
 }