+++ src/openapi/streaming/streaming.ts
@@ -281,7 +281,9 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         this.connection.setConnectionSlowCallback(
             this.onConnectionSlow.bind(this),
         );
-
+        this.connection.setSubscriptionResetCallback(
+            this.onSubscriptionReset.bind(this),
+        );
         // start the connection process
         this.connect();
     }
@@ -558,7 +560,6 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         if (!Array.isArray(updates)) {
             updates = [updates];
         }
-
         for (const update of updates) {
             this.contextMessageCount++;
             this.processUpdate(update);
@@ -680,7 +681,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
                 break;
 
             case OPENAPI_CONTROL_MESSAGE_CONNECTION_HEARTBEAT:
-                // control mesage to keep streaming connection alive
+                // control message to keep streaming connection alive
                 break;
 
             case OPENAPI_CONTROL_MESSAGE_PROBE:
@@ -743,11 +744,16 @@ class Streaming extends MicroEmitter<EmittedEvents> {
 
     /**
      * Resets subscriptions passed
+     * @param subscriptions - subscriptions
+     * @param isServerInitiated - (optional) will be false when we do rest due to missing message  Default = true
      */
-    private resetSubscriptions(subscriptions: Subscription[]) {
+    private resetSubscriptions(
+        subscriptions: Subscription[],
+        isServerInitiated = true,
+    ) {
         for (let i = 0; i < subscriptions.length; i++) {
             const subscription = subscriptions[i];
-            subscription.reset();
+            subscription.reset(isServerInitiated);
         }
     }
 
@@ -971,7 +977,7 @@ class Streaming extends MicroEmitter<EmittedEvents> {
         }
 
         this.connection.onOrphanFound();
-        subscription.reset();
+        subscription.reset(false);
     }
 
     private handleSubscriptionReadyForUnsubscribe(
@@ -1102,6 +1108,10 @@ class Streaming extends MicroEmitter<EmittedEvents> {
 
     private onSubscribeNetworkError() {
         this.connection.onSubscribeNetworkError();
+    }
+
+    private onSubscriptionReset() {
+        this.resetSubscriptions(this.subscriptions, false);
     }
 
     /**
@@ -1250,9 +1260,10 @@ class Streaming extends MicroEmitter<EmittedEvents> {
 
         for (let i = 0; i < this.subscriptions.length; i++) {
             const subscription = this.subscriptions[i];
-            // Reset the subscription and mark it as not having a connection so its state becomes unsubscribed
-            // next action if there is any i.e. subscribe will go in the queue until the connection is available again
-            subscription.reset();
+            // UnsubscribeAndSubscribe will unsubscribe and queue a action to subscribe when we get a result back
+            // we then make the connection unavailable which will stop the subscribe happening until we call
+            // connection available after reconnect
+            subscription.unsubscribeAndSubscribe();
             subscription.onConnectionUnavailable();
         }
 
@@ -1278,10 +1289,11 @@ class Streaming extends MicroEmitter<EmittedEvents> {
 
         for (let i = 0; i < this.subscriptions.length; i++) {
             const subscription = this.subscriptions[i];
-            // disconnecting *should* shut down all subscriptions. We also delete all below.
-            // So mark the subscription as not having a connection and reset it so its state becomes unsubscribed
+            // disconnecting will cause the server to shut down all subscriptions. We also delete all below.
+            // Disposing here causes exceptions if there is any attempt to queue new actions.
             subscription.onConnectionUnavailable();
-            subscription.reset();
+            subscription.reset(false);
+            subscription.dispose();
         }
         this.subscriptions.length = 0;