+++ changelog.md
@@ -1,5 +1,23 @@
+### v10.6.5
+- Fix a bug introduced in 10.6.2 where removing a subscription may cause the snapshot event to still fire
+
+### v10.6.4
+- Do not immediately retry ws auth after a network error
+
+### v10.6.3
+- Reset messages when we get missing messages on the streaming messages
+
+### v10.6.2
+- Fallback to JSON if protobuf errors occur either during the snapshot or an update
+
+### v10.6.1
+- Reduce error logs when it is not a error
+
 ### v10.6.0
 - Tidy up the support for subscribe before streaming setup
+
+### v10.5.1
+- Fallback to JSON if protobuf errors occur either during the snapshot or an update
 
 ### v10.5.0
 - Add support for cloud streaming control message format